+++ Source/ACE.Server/WorldObjects/Player_Inventory.cs
@@ -85,7 +85,7 @@ namespace ACE.Server.WorldObjects
 
             if (packid != 0)
             {
-                wo.ContainerId = (int)packid;
+                wo.ContainerId = packid;
                 AddToInventory(wo);
                 Session.Network.EnqueueSend(new GameMessageCreateObject(wo));
                 if (wo is Container container)
@@ -101,7 +101,7 @@ namespace ACE.Server.WorldObjects
         public WorldObject AddNewItemToInventory(uint weenieClassId)
         {
             var wo = Factories.WorldObjectFactory.CreateNewWorldObject(weenieClassId);
-            wo.ContainerId = (int)Guid.Full;
+            wo.ContainerId = Guid.Full;
             wo.PlacementPosition = 0;
             AddToInventory(wo);
             TrackObject(wo);
@@ -121,7 +121,7 @@ namespace ACE.Server.WorldObjects
             item.PlacementPosition = null;
             item.ContainerId = null;
             // Set fields needed to be wielded.
-            item.WielderId = (int)wielder.Guid.Full;
+            item.WielderId = wielder.Guid.Full;
             item.CurrentWieldedLocation = currentWieldedLocation;
 
             if (wielder is Creature creature)
@@ -163,7 +163,7 @@ namespace ACE.Server.WorldObjects
         {
             EquipMask? oldLocation = item.CurrentWieldedLocation;
 
-            item.ContainerId = (int)container.Guid.Full;
+            item.ContainerId = container.Guid.Full;
             SetInventoryForContainer(item, placement);
 
             RemoveFromWieldedObjects(item.Guid);
@@ -175,7 +175,7 @@ namespace ACE.Server.WorldObjects
             }
 
             // Set the container stuff
-            item.ContainerId = (int)container.Guid.Full;
+            item.ContainerId = container.Guid.Full;
             item.PlacementPosition = placement;
 
             ActionChain inContainerChain = new ActionChain();
@@ -218,7 +218,7 @@ namespace ACE.Server.WorldObjects
         {
             RemoveWorldObjectFromInventory(item.Guid);
 
-            item.ContainerId = (int)container.Guid.Full;
+            item.ContainerId = container.Guid.Full;
             item.PlacementPosition = placement;
 
             container.AddToInventory(item, placement);
